macro fix

+++ hipd/hipd.h
@@ -42,7 +42,7 @@ int hip_agent_filter(struct hip_common *msg);
          {(in6_addr_to)->s6_addr32[0] = 0;                               \
           (in6_addr_to)->s6_addr32[1] = 0;                                \
           (in6_addr_to)->s6_addr32[2] = htonl(0xffff);                    \
-         (in6_addr_to)->s6_addr32[3] = in_addr_from;}
+         (in6_addr_to)->s6_addr32[3] = (uint32_t) (in_addr_from);}
 
 #define IPV6_TO_IPV4_MAP(in6_addr_from,in_addr_to)    \
        { ((uint32_t *) in_addr_to)[0] =                        \
